Split TRX merge engine into focused files
Refactor the merge engine partial class by responsibility while preserving existing behavior.


Copilot-Session: c49afa9d-1e51-4418-bcd0-3914de6c7c5c
